refactor setup
This change is needed because previously, the component was watching the
categories length to see determine when to re-render.

Categories wouldn't change, so there were a few cases where the app
didn't load categories on initial load.

I switched it to watching the profile state, I figure when the user has
his profile, we can consider him logged in, and it's safe to then fire
off the request for categories.