colocate auth/profile selectors with their reducers
This is a pattern we've been following in our reducers which allows us
to place the selector logic in the same place as the reducer logic.

This reduces having to chase down anywhere we may use these selectors in
the event we change our state shape.

A selector used in mapStateToProps should always use state as its first
argument.

If we change our state shape, we need only change the corresponding
selectors in Reducers/index.js:

https://egghead.io/lessons/javascript-redux-colocating-selectors-with-reducers