handle more cases for question manipulation
This is a solution for handling more cases that come up when considering
seamlessly handling questions that may only exist in state (those from
opentdb which a user hasn't decided to save yet), and those that did
come from our database.

Most cases should be accounted for at this point.