add Setup component back into root
this component will ensure that a user is properly logged in and has
their categories and questionTypes, and that these requests are made in
order.

we also needed to adjust our auth logic since we switched it to using
emails for user creation on the backend.

the action that should be taken after we retrieve an auth0 token should
always be:
1.) Fetch profile from auth0.
2.) Get email from that profile and POST it to our backend no matter if
  we've seen the user or not.